+++ ResourceAccess/FitnessApp.Core.ResourceAccess/Models/WorkoutItemModel.cs
@@ -17,7 +17,7 @@ namespace FitnessApp.Core.ResourceAccess.Models
 
         [Required]
         [StrictlyPositiveProperty("distance")]
-        public double Distance { get; set; }
+        public int Distance { get; set; }
 
         [StrictlyPositiveProperty("calories")]
         public int Calories { get; set; }
